requirements.txt
asgiref==3.11.1
Django==6.0.3
django-rest-framework==0.1.0
djangorestframework==3.17.1
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
